Correcting moving values

+++ app/src/main/java/fr/iutlens/mmi/demo/game/sprite/sprites/Projectile.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.launch
 import kotlin.math.abs
 import kotlin.math.round
 
-class Projectile(var sprite: BasicSprite, val friendly : Boolean = false, var speed: Float, var range: Float, var damages: Int, var knockback : Float, var onHitEffects : MutableList<(character : Character)->Unit> = mutableListOf()) {
+class Projectile(var sprite: BasicSprite, val friendly : Boolean = false, var speed: Float, var range: Float, var damages: Float, var knockback : Float, var onHitEffects : MutableList<(character : Character)->Unit> = mutableListOf()) {
 
     fun realSpeed(game: Game) : Float{
         return speed*((game.map.tileArea.w + game.map.tileArea.h)/2)
